backend/requirements.txt
Django==5.0.1
djangorestframework==3.14.0
django-cors-headers==4.3.1
PyMySQL==1.1.0
python-decouple==3.8
djangorestframework-simplejwt==5.3.0
setuptools>=65.0.0
